commit 商城后台 可以上传图片

+++ management_mms_product_consumer_xxl/src/main/java/com/cebbank/management_mms_product_consumer/common/pojo/ProductDetail.java
@@ -45,7 +45,7 @@ public class ProductDetail extends Model<ProductDetail> {
     /**
      * 商品价格
      */
-    private BigDecimal productDetailPrice;
+    private Double productDetailPrice;
 
     /**
      * 商品库存
